Update Stream_First_Non-repeating.java
Given an input stream s consisting only of lowercase alphabets. While reading characters from the stream, 
you have to tell which character has appeared only once in the stream upto that point. 
If there are many characters that have appeared only once, you have to tell which one of them was the first one to appear. 
If there is no such character then append '#' to the answer.